// frontend/src/types/dashboard/adminDashboard.ts
export interface AdminDashboardData {
  totalCategories: number;
  totalSubCategories: number;
  totalUsers: number;
  totalCustomers: number;
  totalVendors: number;
  totalVendorStores: number;
  totalPendingVendorStores: number;
  totalApprovedVendorStores: number;
  totalRejectedVendorStores: number;
  totalProducts: number;
  totalPendingProducts: number;
  totalApprovedProducts: number;
  totalRejectedProducts: number;
}